[dnssd] Replace invalid forced reset logic (#24698)
* [dnssd] Replace invalid forced reset logic

The platform implementation of DNS-SD contains a mechanism
for restarting the advertising when the platform returns
the CHIP_ERROR_FORCED_RESET error. This mechanism is broken
as it assumes that only one operational service is used and
it uses extra RAM although it is only used by Linux.

Switch to an event-based approach that allows the DNS-SD
server to restart the advertising when needed.

* [dnssd] Make sure initialization is not run twice

Currently, the platform implementation of DNS-SD uses
a boolean member to indicate if the backend has already been
initialized, but the initialization is asynchronous so it
may happen that the initialization is invoked twice. Add
another "initializing" state to handle that case.

* Improve comment

* Fix build